feat: Add sitemap generation for blog posts
This commit adds the necessary code changes to generate a sitemap for the blog posts. It introduces a new file, `sitemap.xml.ts`, in the `server/routes` directory, which uses the `sitemap` package to generate the sitemap. The URLs of the blog posts are added to the sitemap with a monthly change frequency. This feature improves the SEO of the application and helps search engines discover and index the blog posts.